fix: Don't rely on response's structure
For an unknown reason, the responses from Macaulay Library's API are not well
formed: the `count` item aggregating the number of responses to the query made
is always zero. See below:

```
{
  "results" : {
    "count" : 0,
    "content" : [ {
      "specimenUrl" : "https://macaulaylibrary.org/asset/148672301",
      "catalogId" : "148672301",

    [...response clipped...]
```

This change relies in the result's content's array length instead of
`results.count`, preventing the service falling on a retrial loop.